UI changes required

+++ admin/Category-List-delete.php
@@ -7,7 +7,7 @@ if(!isset($_REQUEST['id'])) {
 	exit;
 } else {
 	// Check the id is valid or not
-	$statement = $pdo->prepare("SELECT * FROM tbl_end_category WHERE ecat_id=?");
+	$statement = $pdo->prepare("SELECT * FROM tbl_category_type WHERE ctype_id=?");
 	$statement->execute(array($_REQUEST['id']));
 	$total = $statement->rowCount();
 	if( $total == 0 ) {
@@ -18,58 +18,76 @@ if(!isset($_REQUEST['id'])) {
 ?>
 
 <?php
-	
 
 	// Getting all ecat ids
-	$statement = $pdo->prepare("SELECT * FROM tbl_product WHERE ecat_id=?");
+	$statement = $pdo->prepare("SELECT * FROM tbl_category WHERE ctype_id=?");
 	$statement->execute(array($_REQUEST['id']));
+	$total = $statement->rowCount();
 	$result = $statement->fetchAll(PDO::FETCH_ASSOC);							
 	foreach ($result as $row) {
-		$p_ids[] = $row['p_id'];
+		$cat_ids[] = $row['cat_id'];
 	}
 
+	if(isset($cat_ids)) {
 
-	for($i=0;$i<count($p_ids);$i++) {
-
-		// Getting photo ID to unlink from folder
-		$statement = $pdo->prepare("SELECT * FROM tbl_product WHERE p_id=?");
-		$statement->execute(array($p_ids[$i]));
-		$result = $statement->fetchAll(PDO::FETCH_ASSOC);							
-		foreach ($result as $row) {
-			$p_featured_photo = $row['p_featured_photo'];
-			unlink('../assets/uploads/'.$p_featured_photo);
+		for($i=0;$i<count($cat_ids);$i++) {
+			$statement = $pdo->prepare("SELECT * FROM tbl_product WHERE cat_id=?");
+			$statement->execute(array($cat_ids[$i]));
+			$result = $statement->fetchAll(PDO::FETCH_ASSOC);							
+			foreach ($result as $row) {
+				$p_ids[] = $row['p_id'];
+			}
 		}
 
-		// Getting other photo ID to unlink from folder
-		$statement = $pdo->prepare("SELECT * FROM tbl_product_photo WHERE p_id=?");
-		$statement->execute(array($p_ids[$i]));
-		$result = $statement->fetchAll(PDO::FETCH_ASSOC);							
-		foreach ($result as $row) {
-			$photo = $row['photo'];
-			unlink('../assets/uploads/product_photos/'.$photo);
+		for($i=0;$i<count($p_ids);$i++) {
+
+			// Getting photo ID to unlink from folder
+			$statement = $pdo->prepare("SELECT * FROM tbl_product WHERE p_id=?");
+			$statement->execute(array($p_ids[$i]));
+			$result = $statement->fetchAll(PDO::FETCH_ASSOC);							
+			foreach ($result as $row) {
+				$p_featured_photo = $row['p_featured_photo'];
+				unlink('../assets/uploads/'.$p_featured_photo);
+			}
+
+			// Getting other photo ID to unlink from folder
+			$statement = $pdo->prepare("SELECT * FROM tbl_product_photo WHERE p_id=?");
+			$statement->execute(array($p_ids[$i]));
+			$result = $statement->fetchAll(PDO::FETCH_ASSOC);							
+			foreach ($result as $row) {
+				$photo = $row['photo'];
+				unlink('../assets/uploads/product_photos/'.$photo);
+			}
+
+			// Delete from tbl_photo
+			$statement = $pdo->prepare("DELETE FROM tbl_product WHERE p_id=?");
+			$statement->execute(array($p_ids[$i]));
+
+			// Delete from tbl_product_photo
+			$statement = $pdo->prepare("DELETE FROM tbl_product_photo WHERE p_id=?");
+			$statement->execute(array($p_ids[$i]));
+
+			// Delete from tbl_product_size
+			// $statement = $pdo->prepare("DELETE FROM tbl_product_size WHERE p_id=?");
+			// $statement->execute(array($p_ids[$i]));
+
+			// Delete from tbl_rating
+			$statement = $pdo->prepare("DELETE FROM tbl_rating WHERE p_id=?");
+			$statement->execute(array($p_ids[$i]));
+
 		}
 
-		// Delete from tbl_photo
-		$statement = $pdo->prepare("DELETE FROM tbl_product WHERE p_id=?");
-		$statement->execute(array($p_ids[$i]));
-
-		// Delete from tbl_product_photo
-		$statement = $pdo->prepare("DELETE FROM tbl_product_photo WHERE p_id=?");
-		$statement->execute(array($p_ids[$i]));
-
-		// Delete from tbl_product_size
-		// $statement = $pdo->prepare("DELETE FROM tbl_product_size WHERE p_id=?");
-		// $statement->execute(array($p_ids[$i]));
-
-		// Delete from tbl_rating
-		$statement = $pdo->prepare("DELETE FROM tbl_rating WHERE p_id=?");
-		$statement->execute(array($p_ids[$i]));
+		// Delete from tbl_category
+		for($i=0;$i<count($cat_ids);$i++) {
+			$statement = $pdo->prepare("DELETE FROM tbl_category WHERE cat_id=?");
+			$statement->execute(array($cat_ids[$i]));
+		}
 
 	}
 
-	// Delete from tbl_end_category
-	$statement = $pdo->prepare("DELETE FROM tbl_end_category WHERE ecat_id=?");
+	// Delete from tbl_category_type
+	$statement = $pdo->prepare("DELETE FROM tbl_category_type WHERE ctype_id=?");
 	$statement->execute(array($_REQUEST['id']));
 
-	header('location: Category-List.php');
+	header('location: category.php');
 ?>
